fix: remove uploader from list when upload is successful
Part of story #38627: Choice of a new Rich Text Editor

What has been fixed:

When you upload a file in a prosemirror editor, and you save the
content of the section, a call is made to DELETE uploads/tracker/file
because the upload file plugin is trying to abort all the ongoing file
uploads. It results a failed request with a 404 error code, because the
file upload does not longer exist backend-side.

Spotted while fixing artidoc's Cypress tests.

Removing the uploader from the list of uploaders after its done
fixes the issue.

How to test:
- Pick up an editor
- Drop a file in this editor
- Open your web console and go to the "network" pane
- Save the section

--> No DELETE call has been made.

Change-Id: I9b0545be94fd289cae26f104ac84ebdda636653d