fix: popover elements not found when toolbar is disconnected
Part of story #38627: Choice of a new Rich Text Editor

Context:

The popovers of the link and the image buttons are moved to
document.body right after the toolbar has been connected to the DOM (see
git #tuleap/stable/002cb45ebe1d2abb3520efa3d31a3c397bfaacdc) and are
moved back to the toolbar element when it is disconnected from the DOM.

However, we never had a case where the toolbar was removed from the DOM,
and thus never caught this issue. It was spotted while fixing artidoc's
Cypress tests in which the error "Unable to find the popover element"
error was logged twice in the tests logs.

How to test:
- Run the cypress tests of artidoc.
--> There are no "Unable to find the popover element" errors anymore.

Change-Id: I4769b50af3540dd4ed7ca6dca9fd89bdb33a426a