fix: unable to upload file after adding the field
Given I created artidoc sections on a tracker without attachment field,
when I finally add the field because my teammates want to include funny
pictures of kittens and puppies in their requirements document, then all
existing sections cannot have attachments.

Since the artifacts were created before the attachment field, the
corresponding field value is null for the last changeset, leading to a
null REST field value. The fix to the issue is to return an empty REST
value in this case.

To test:
* Create a tracker without attachment field
* Create an artidoc based on this tracker
* Add a section, save it, reload the page
* => you should be able to edit the section and upload a picture of a
  kitten or a puppy.
* => you should be able to add a new section and upload a picture of a
  kitten or a puppy.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: I01183051da6da8d6e50d4865b87c8cd80f1d1569