fix: request #40113: Artidoc icon in tooltip
When you reference an artidoc document, the tooltip should display the
icon of artidoc, not the generic "binary" one.

Likewise, when an artidoc is referencing (via its document/docman
description for example) an artifact, then the reference displayed in
cross reference field in the artifact view should display the right
icon.

Change-Id: Ib07890eca254f531496819861384cd6ef8f4209d