fix: loadTooltip only on current description
When an artidoc has many sections and the first section contains a
cross reference (to an artifact for example) then the hover on the cross
reference should trigger only one network call to fetch the
corresponding tooltip (`goto` url).

(Note: cross reference tooltips are only available when you don't have
write permissions on the document)

When we call `loadTooltips` without any argument, it will load tooltips
on the whole document, meaning that the cross reference of the first
section will be loaded many times.

Part of story #40114: Cross references in Artidoc

Change-Id: Idf856513b020f4fdcdd3ae79db3fd4ae787ed309