fix: dropdown menu should not appear below the TOC
The menu of the section dropdown is moved to the document.body in order
to be sure that it will be positioned above the TOC. For reference:
git #tuleap/stable/002cb45ebe1d2abb3520efa3d31a3c397bfaacdc

This works well for section and the add of existing section. However
there is a timing issue for pending issue. A pending issue does not have
any dropdown (there is no artifact to go to and the removal is done with
the cancel button). The dropdown is displayed as soon as the section is
saved (and not anymore pending). This means that we must move the menu
when the dropdown is initialized instead of when the component is
mounted.

To test, add a new section, save, open the dropdown menu => it should
appear above the TOC.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: Ie420328cba0cc822327c0004a595a59b7a682300