feat: Remove useless Edit option in dropdown
The dropdown of the section doesn't need anymore an Edit option since
the section is always in edit mode thanks to prosemirror.

Took the opportunity to reword the "Delete" option to "Remove" so that
users don't think that the artifact will be deleted.

Part of story #40114: Cross references in Artidoc

Change-Id: Ia3c88aa887db5741c472c37965d8f3faed3cba2f