refactor: split tooltip.ts
tooltip.ts is big ball of different functions used to instantiate,
fetch, and display tooltips on cross references.

In order to be able to update the behavior of this library, this
contribution split it in different files so that it will be easier to
add tests.

Some tests have been added as a best effort, the others are left for the
future us.

No functional changes, tooltip is still displayed on cross references,
on user @ mentions, and in roadmap.

Part of story #40114: Cross references in Artidoc

Change-Id: I7bee220152c2a7c6fde27c3d101028f5028979b5