fix: alignment of artidoc icon in xref
When a document references an artifact (for example in its document
description), then in artifact view there is an entry in the cross
reference field with the title of the document and an icon.

The issue is that if the document is an artidoc, then the icon is not
centered in the enclosing disc. Update of the design (paths & anchors)
of the icon in TLP font fixes the issue.

No functional changes for other icons in the font (TLP doc » visual
assets » iconography).
No functional changes in the display of artidoc icon in tree view.

Discovered in git #tuleap/stable/65a87bf268f686079682e206c11c81a442327501

Part of request #40113: Artidoc icon in tooltip

Change-Id: If86d9d0581b996e080f0d2167e123b88d8690967