feat: Expert query in placeholder
part of story #40088 Expert mode is the new default for cross tracker
search

How to test:
- Add a new cross-tracker search widget (in expert mode since it's new)
- Edit the query, clear the editor.
- The placeholder text should be a query valid for expert mode, not a
query for default mode.

Notes:
If you switch the mode to default immediately, the placeholder does not
change. If you switch to default mode, save the report, and then clear
the editor, it will be the placeholder for default mode. I'm not sure we
need to go that far to fix this, the main goal of this change is to NOT
expose end-users immediately to a query that cannot work because it's
missing SELECT FROM WHERE. We will need to tinker with CodeMirror too
and already plan to on a separate patch.

Change-Id: Ie706463d5630221a2c3000e535004dcdcaafe34e