feat: Add Quick look access to recent items
If a document link appears in the recent items in SwitchTo, then the
main action is to follow the link. This means that if the user wants to
interact with the docman item (set approval table for example) there is
no way to access to the document, they must go through project »
documents » find the document.

This contribution adds a quick link to be able to reach the Quick look
view of the document.

Part of story #39358: Artidoc in SwitchTo

Change-Id: I1cd897035b9d8d455dcc107a12c36b0932386611