fix: forbid marks in artidoc-section-title
part of story #38627: Choice of a new Rich Text Editor

The title node is supposed to contain only text, but it turns out that
marks are still allowed inside of it. The user is then able to put the
title in italic using the keyboard shortcuts.

Setting explicitely `marks: ""` in the artidoc_section_title node spec
disables the usage of marks inside the node.

How to test:
- select some text in the title
- press `ctrl+i` or any key binding setting a mark on a text node
--> Nothing happens

Change-Id: I91400f24a354177e84bc761b50369362a9f12e49