fix: unable to set text style in Chrome
part of story #38627: Choice of a new Rich Text Editor

As it turns out, click handlers set on HTMLOptionElements will never be
called in Chrome based browsers. Since the options in the text-style selectbox listen for clicks, then the text styles are
never applied. Let's put a onchange handler on the select element instead.

How to test:
--> Both in Chrome and Firefox, setting text style should work as
expected.

Change-Id: I643773ed77a34209d0db7977981fb458fb24ff64