feat: artidoc as document
Since the aim of an artidoc is to work on a document based on artifacts,
representing this document as a real life document will better match
user expectation about a document editor (most of popular document
editors display such documents in a similar manner).

This has the benefit to not have lengthy lines displayed on wide
screens so it will provide a better reading experience.

Closes request #40437: artidoc as document

Change-Id: I8ac56de7f0a2ef104609acabbdea040abdd1828c