feat: Retrieve widget configuration
Part of story #37941 Timetracking management widget

Retrieving widget configuration from database
When a configuration is saved ("Save query" button) it is saved in the database and will appear each time the page is reloaded.

The query is injected directly into the template, so there's no GET to the REST API.

How to test:
-- Click on dates to switch to edit mode, change something, then click on the "Search" and "Save query" buttons
-> Your changes should appear on the widget and in the database.
-- Reload the page
-> Your changes should still be displayed in the widget

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

Change-Id: I6411af39e153e82bf3919b4a78c69efbb2b85d8f