feat: Heading 1 becomes Large heading
When we will have structured artidoc, sections will be nested inside
sections. This means that there will be different levels of "Heading 1".
This might be confusing to the writer to put imbricated "Heading 1"
titles, we think that "Large heading" will be better.

Took the opportunity to rename "Texte" style to "Normal" to better match
user expectations:
* LibreOffice: "Text body"
* OnlyOffice: "Normal"
* Google Drive: "Normal text"
* Office 365: "Normal"

part of story #38627: Choice of a new Rich Text Editor

Change-Id: I8e37840daae48c2cad30a034bf1905f3c4fc70f8