chore: Composition API into CreateGitlabLinkModal
Part of request #31168 Migration Vue2 => Vue3

Add Composition API into git/repositories-list -> CreateGitlabLinkModal

Why?
Class components are no longer supported in Vue 3. We must switch to composition API to upgrade to Vue 3.

Note about `getGitlabRepositoriesIntegrated` in `ListRepositoriesModal.vue`:
Most tests of this component failed due to the error `Vue 2 does not support readonly arrays`.
The `useGetters` encapsulates getters in a readonly property, but Vue 2 does not support readonly.
The only solution that works is to access the getter directly via the store to avoid the readonly property.

This solution is not intended to stay as it is, and will be modified as soon as we switch to Vue3.

No functional changes expected

Change-Id: Iaece36956d1bab3596c9bf18f994161556fee69d