feat: remove at mention autocompletion from text fields
part of story #38631 @ notifications in trackers

How to test:
- Create a new artifact.
- Edit a text field. When you type "@", there is no autocompletion of
usernames
- There is no "@" autocompleter with Text, HTML or Markdown formats.
Change your user's default format in User preferences to test.
- There is no "@" autocompleter in String fields.
- Edit an existing artifact.
- Write a new follow-up comment. When you type "@", there should be
autocompletion of usernames. It should work with all formats (Text, HTML
or Markdown).
- No other functional change for text fields or follow-up comments.
- In "Mass change" view, there is no longer autocompletion in Text and
String fields. Behaviour for the follow-up comment is unchanged.
- In "Copy artifact" view, there is no longer autocompletion in Text and
String fields.

Notes
I grouped 10 of the parameters of "getTextarea()" in a "DTO". Since many
of those parameters vary together, it falls to only 5 parameters to
build a Configuration object. The "is artifact copy" parameter is kept
separate because it is exclusively related to image upload support, and
this is computed in the Provider.
I also updated the baseline because I must use a deprecated method
"setCacheSpecificProperties" in the Text field builder to set field
properties like the number of rows and the number of columns.

Change-Id: I69e8bc7094a6d58028822ea45ccd94aada2dd36e