refactor: presenter for RichTextarea provider
part of story #38631 @ notifications in trackers

No functional change expected in Artifact view: when image upload is
enabled (there is a File field and user can update it), it still
functions both in Text field and in the New follow-up comment.
To test image upload, switch to HTML format and either drag/drop an
image or copy/paste it in the editor.

Why?
We will work in this area to remove the @mention support for Text
fields, so we might as well have a proper presenter.

The files are moved in accordance with ADR-0030 PHP folder structure for
plugins.

Change-Id: Ic0e2369831a42449124e0087e57fe2ec4dfd729f