chore: build modal v2 separately
part of request #29264 Split the build of tracker plugin frontend assets
into multiple apps

How to test:
No functional change in Cardwall view or Tracker report Cardwall
renderer. The modal v2 still loads and opens as before when you click on
"Create a task" in Tracker Cardwall renderer or when you edit the
Release in Cardwall.

Change-Id: I1faec7c53e158eeba50e94d4d69cc021859a2fe4