chore: Apply PSR-4 for Tracker unit tests
part of request #17148 Tracker move to PSR4 😱

No functional change. Unit tests should pass.

Notes:
Changed the limit to switch between "running only on modified files" and
"running on the whole codebase" from 1000 modified files to 17266
characters. It seems this patch hits a limit of character count when
passing the list of 900+ modified file paths to the docker command to
run PHPCS. The limit does not evoke a power of two, but was guessed
based on where the Jenkins log started cutting off the list of files.

Some stubs were actually dead code and have been removed.
The baseline has been updated in place, this commit does not attempt to
fix Psalm issues.
The files are moved in accordance with ADR-0030 PHP folder structure for
plugins.
require_once calls to bootstrap.php have been removed (they are not
needed by PHPUnit).

Change-Id: If2acb8b94b4ea1758687fd7528e1b10018cc9663