fix: chrome should indicate that user click on an image
Given an artidoc section contains an image,
When user clicks on it,
Then the browser should indicate that the image is selected (for example
to be able to delete it)

On Chrome browsers, nothing happen.
On Firefox, the image has an overlay.

I tried to force the overlay for everyone with the following:
```css
img::selection {
  background: var(--tlp-main-color-transparent-50);
}
```

However still nothing happen on Chrome and I didn't find why.

As a workaround, we use the ProseMirror specific class that is added
when we click on an image: `ProseMirror-selectednode` and add a border.

Note: on different environments than mine, it appears that Chrome is displaying an overlay.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: Ic39dfad3a7961404e92fb9d603d055edf385e251