fix: Do not incitate user to drop when not allowed
If a section is based on a tracker without attachment field, then user
should not think that images can be dropped in the editor area.

The dropcursor is disabled, and if the user still drop an image, then an
error message is displayed.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: I058707c12ee0a76d92e92b69c51344a3e1b8cb1c