chore: fix Psalm errors in Stubs & Builders
part of request #17148 Tracker move to PSR4 😱

No functional change. Unit tests and Psalm should pass.

Notes:
Psalm was reporting a type mismatch error for a call to
ChangesetTestBuilder::aChangeset(), so I chose to change the type
everywhere to int (as it should be). When I wrote the builder, I used
string because I had been burned by the DAO returning a string id. We
should instead set-up our builders so they build correct data, the id
should be int.

Change-Id: Iaa39365bdac0b37fc0bfb1785ca3559d821bad3c