chore(forge-config): Transfer local.inc.dist to class const
The goal of this commit is to get rid of `src/etc/local.inc.dist` to
limit the number of possible sources for configuration variable
definitions.

The tradeoff here is to import the variables into a dedicated class to
reduce the number of changes in the source code (compared to replace
usage of strings by usage of const everywhere). Some variables already
had a double existence, as class const and in local.inc.dist for default
value, in these case, the transfer and adjustment were done.

Part of: request #40460 Have a single source of truth for configuration variables

Change-Id: I34c8e3ec39e3afb536bdee8499250fa02d133fc6