refactor: artidoc containers height
Rewrite CSS rules to have TOC that has a full visible height (the issue
is visible when there is too many items in the TOC).

No functional changes
* In small screens
* In medium screens
* In large screens
* With or without project background
* When permission denied to the tracker
* When document is empty (and user not writer)
* When document is just created (config screen)
* When project toolbar is displayed
* When platform toolbar is displayed

In order to have a reliable computation for the height of the TOC, we
cannot anymore stick the toolbar to the navbar header on scroll. The
header is not anymore pinned: breadcrumbs and document title are always
visible => no need anymore to scroll to the top of the page to print the
document or to access the parent folder of the current document.

Note: starting from this contribution, the section title (and its
number) is not anymore sticky:
* This is not the case when we edit an office document
* When we will have sections of sections the choice to know which title
  to stick will be challenging. I suggest to defer the decision for
  later.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: Iaa6e7b8e125190fb3ba2dd345a9c15214006ea06