chore: build tracker styles separately
closes request #29264 Split the build of tracker plugin frontend assets
into multiple apps

How to test:
No functional change:
- Tracker FlamingParrot and BurningParrot pages should still load their
CSS styles.
- In Tracker Admin > Workflow > Field Dependencies, the CSS style of the
matrix table should still be loaded.
- CSS styles for Print mode (add "&pv=1" in Artifact view) should still
be loaded.

Notes:
While we use CSSViteAssets in the event handler for cssFile(), we cannot
use Layout::addCssAsset because it is too late in the execution. It
looks like the HTML for the header has already been sent, using
addCssAsset at this point has no result. Improving the loading of the
CSS (by not using global events for a start) should be a separate
effort, the situation should not be worse with this patch. Note that due
to the API, we are forced to build a useless ThemeVariation. It is
discarded by CSSViteAsset.

This also corrects a mistake I made when setting up vite configurations
in Tracker. Due to ad-blockers filtering "assets/tracker/", we have used
a workaround "assets/trackers/" public path for Tracker. It means we
cannot use the basename of the plugin folder, so we just use "trackers"
magic string. It has not caused problems until now because we were not
using dynamic import, but CSS styles import some loader images, and they
will fail to load without this. I took the opportunity to fix the
mistake everywhere.

There are still images remaining in themes/default/images/ folder. This
is intended: they are used in an email template. I did not want to try
and see what happens if we move them: they are never used in CSS.

Change-Id: I56781e7385645182ac228f42245a03feb1840622