feat: collapse sidebar
The TOC can now be open or closed in medium and large screens.

In small screens the TOC is always displayed, because it is always
displayed on top of the document and do not take horizontal place, so no
need to open/close it.

In medium screens if there is enough room, when sidebar is collapsed the
artidoc is displayed as a document (when expanded, artidoc still takes
the full width).

Note: toc/ will be moved to sidebar/toc/ in a dedicated contribution to
not clutter the review.

Part of story #40457: Modify flat document outline

Change-Id: If49d3d55fd2ffcc5321f0ba1efe55a6acc055373