Cursor can go back into title when user press backspace
part of story #38627: Choice of a new Rich Text Editor

How to reproduce:
have a simple text in artidoc
```
test 1
```

add a space at the end of sentance, then remove it.

The cursor is now placed in title section

Why?
When user click on `Cancel` button, we do want to reset editor content, we have a watch over prop `is_edit_mode`
when we go back to original description we trigger the `is_dit_mode` watch
we should check if the editor content is the same,
we want to reset full editor content only when user have click on `Cancel`

Change-Id: I28d52b7d8b9ce0d7133c03067a6b79c779fe1a69