refacto: Modernize Masschangeupdater class
Part of story #38631 @ notifications in trackers

This change will make easier to handle the notification send after the
updates of the artifacts in the mass change. A new paramater will be
introduced in PostCreationContext to handle the users mentioned in the
follow up comment.
Also, Artifact::createNewChangeset is deperacted, so, it is also an
opportunity to replace it with NewChangesetCreator::create in the mass change case.

Some mocks has also been removed in the test

No functional change expected in the mass change, the update and post
action update should work as before.

Change-Id: I953035ccde437792d02d2126ff281c753879419e