fix: Bug with right arrow after pressing enter from title
Part of story #38627: Choice of a new Rich Text Editor

What has been fixed:
When you have a section with a title and at least 2 rows in the description.
Click on the title and press 'Enter'. You are now at the beginning of the description.
Press "ArrowRight" to move to the second position of the description.
The cursor moves to the second row instead of after the first character.

Change-Id: I38f23cdc33f69d13356324eb1ab07c8be3cbd9de