feat: design dragndrop grip in artidoc TOC
This is only a cosmetic contribution with a focus only on the design of
the dragndrop grip according to mockups.

Initial plan was to use fa-grip-vertical, but dealing with font-size
in this area easily melt my brain therefore the svg from the mockup has
been used instead.

Part of story #40457: Modify flat document outline

Change-Id: If812dfe5718937db9afb8e773fb2bc54d75a1a15