chore: Move graphontrackers script to Vite
Part of request #40462 Modernize GraphOnTrackersV5

According to ADR 0018, we should stop use webpack and use vite instead.
Migrate the script of the plugin to vite build system.

Note: Reports rendered are loaded old way by echo-ing html directly and
assets are included after. With vite assets containing css we need to
include them before the header. In order to achieve that I added a
method to retrieve the assets of the renderer and add them directly to
layout.

*No functional changes expected*

Graphs in tracker reports and in widgets should continue to work as
before

Change-Id: I1a40c75629bb2576813db0d46f787388e3a5e2ce