chore: improve ability to replay tracker e2e tests
part of story #38631 @ notifications in trackers

No functional change. make tests-e2e should pass.

Why?
We want to add more e2e tests related to artifact and mass-change. It
seems the tracker_artifact suite is the place for them. Before starting
however, we should improve the ability to replay those tests so that we
can work comfortably.

All tests in tracker_artifact test suite depended upon an XML imported
project. For all tests where it was straightforward, this is changed: we
now create a new project before all tests that are not for siteadmin. In
some tests, the test depended upon the structure of the tracker, so we
cannot "just" use the Issues template, for example the "computed field"
test requires a computed field, the "Move artifact" test requires a
certain structure. It would be nice to change those: maybe the test
should add a field, etc. I chose to not deal with them yet and do
"better than before".

The tests are also changed to never depend on dynamic data-test
attributes: we should never have introduced those in the first place.
Instead, we use static data-test (or sometimes with a suffix to avoid
erasing the existing data-test attribute) and we check what the user can
see: the label of the field.

I also changed the tests to go faster: whenever possible, we visit()
directly to the right page, we remove useless visits to the Trackers
service, etc.

Change-Id: I0dcc171b5eeabfda133b067a91d98d92ea5152c9