feat: design reorder arrows in artidoc TOC
This is only a cosmetic contribution with a focus only on the design of
the reorder arrows according to mockups.

Overflow for arrows is applied in the same manner than for planning
view for consistency.

User can access to arrows with the keyboard (traditional `tab` until the
focus is on the arrow.

Part of story #40457: Modify flat document outline

Change-Id: If771415672d0d31a3785f98e5f6c8111003786bd