fix: request #40465 Should show errors when renaming TTM campaign
While editing the label of a TTM campaign, if there was an error like a
too long name, the error was not shown. Instead we just get nothing and
the spinner spinning.

The same problem appear in "Configure automated tests" modal

*Testing:*

You can force a RestException to be thrown in CampaignsResource::patch()
or edit the label of the campaign to a name longer than 150 characters

-> An alert should be shown displaying the error message

Change-Id: I668b7a8555d601f907670d040da1854bb05a76c7