feat: sections can be moved up/down
When user clicks on up/down arrows in the TOC, then sections are
reordered accordingly. Note: this is not saved since backend is not
ready yet to persist the new order.

When user use the keyboard (enter on the up/down button), then the focus
is kept on the button so that she can continue to move the section up or
down.

Part of story #40457: Modify flat document outline

Change-Id: Ie8cacd83800b7d64c3299a215c8f0d6f771347f4