chore: end-to-end test for edition of comment
part of story #38631 @ notifications in trackers

No functional change. make tests-e2e should pass.

Why?
We are going to overhaul the edition of comments. It is currently
written in PrototypeJS, we will replace it. Thus, having an end-to-end
test that covers this part will let us work more confidently, even if it
just tests the "Happy path".

empty_followup has been renamed to "private_comments", because this is
the actual feature that it tests. I also replaced the dynamic data-test
attributes. Only static ones are used now for follow-ups.

Change-Id: I814ab82299ce23d2818ddc9508687640de7e6747