fix: add new section buttons position with collapsed TOC
On small screen size, the TOC should be displayed above the artidoc.
The document takes all the available width.
Add new section button is displayed inside the artidoc. (no changes)

On medium screen size, the TOC should be displayed in the right sidebar.
If this sidebar is expanded, then the artidoc takes all the available
width.
Add new section button is displayed inside the artidoc. (no changes)

If the sidebar is collapsed, then if there is enough room, then the
artidoc will be narowed so that it is displayed as a document.
In that case, the Add new section button should be displayed outside of
the artidoc, in the left margin. (This is the purpose of this
contribution)

On large screen size, the TOC is still displayed in the right sidebar.
Sidebar collapsed or not, the artidoc is displayed as a document.
Add new sectino button is always displayed outside of the artidoc, in
the left margin. (no changes)

Part of story #40457: Modify flat document outline

Change-Id: I0a6fdd855fd665352f1792aa3a14fe2c7a0bc084