chore: trash current cross-references integration
part of story #40114 Have better Cross References in Artidoc

The integration of cross-references in our ProseMirror editor is about
to be redesigned from scratch with a completely different approach.

Let's trash the current implementation and cleanup the code to start on
a new base.

Note 1: I kept the project_id in the useEditor function because it
will be needed in the next contribution.

Note 2: Although there are no more cross-reference links to update in
the editors, I chose to keep the implementation of the cross-reference
link popovers. It will be reused once we'll have them back in a couple
of commits.

How to test:
--> No references are loaded, detected nor displayed.

Change-Id: Ie29504a6b83ef8ba38929b080a133c356cf21d55