feat: detect typed cross-references
part of story #40114 Have better Cross References in Artidoc

Implementation notes:

1. The custom mark "async_cross_references" allows us to flag pieces
of text as cross references. It takes two attributes:
1. `text` the text matching the format of a Tuleap cross-reference
2. `project_id` the id of the current project

This mark has no defined `parseDOM` attribute because we will never have
to rebuild this mark from existing DOM (unlike the other marks). We
explicitely prevent it to be serialized, so it does not end up in
the section's HTML description. In the editor view, it is rendered as an
AsyncCrossReference custom element.

2. The AsyncCrossReference element's role is query the backend to know
   whether the reference text actually matches an existing reference.
When it is the case, it sets its own data-href attribute with the found
reference's url, and loads its own tooltip. When the text is modified
inside of the custom element, the custom elements updates itself. When
no reference is found, it will stop displaying itself as a valid
cross-reference.

3. The input rule DetectCrossReferenceAsYouType's role is to insert the
   custom mark when the text typed right before the cusor matches the
expected format. Please note that the regexp is not perfect, and might
be reworked in a dedicated patch.

4. The PluginInput has been reworked. It now requires a DOMSerializer
   (to avoid instanciating each time the document is updated), and no
longer execute its callback in the view.update hook. Instead, it reacts
only when the document has changed, meaning that it will execute the
callback only when significant things have been done in the document
(formatting, typing, deletions etc.).

How to test:
- In an editor, type an existing cross reference followed by the space
  character.
--> It is displayed as a link
--> Its tooltip appears when hovered
--> Its popover menu appears when clicked

Note that it is expected that nothing happens if you try to update it
through the popover menu since no plugin react to the
"updated-cross-reference" transaction yet.

- Type a non existing cross reference followed by the space character.
--> It is not displayed as a link

Change-Id: Ia50d8c07720a7f8f8e724f6ea3ef9e8e9e7d72e7