refactor: kill prototype in comment edition
part of story #38631 @ notifications in trackers

No functional change. make tests-e2e should pass (on tracker_artifact
suite at least).

Note that the textarea or editor is not focused after opening it, this
was already the case on master.

Suggested entry point for the review:
plugins/tracker/scripts/artifact/src/edition/edit-view.ts

Why?
Prototype framework is end-of-life since almost 10 years. We need to
replace it by something. It's not worth it to pull Vue.js for such a
small part of the artifact view, therefore I chose to go with Vanilla
TypeScript + lit-html.
The textarea is not rendered by lit-html because the Rich Text Editor
lib needs it to wrap it (or not) with a CKEditor. It was simpler to
"just" create it via Vanilla DOM and pass it to the lib, we don't need
to depend on a special ID format anymore. The RTE-creator lib has been
adapted to return the created editor, this way we can call
`getContent()` instead of using the global variable.
I had to remove the `isAjax()` check because we are now using fetch, not
XHR.
The code now uses data-attributes instead of CSS classes, but I did not
rename the classes as most of them are still used by other legacy
scripts.
I tried to move most of the CSS classes dedicated to comments in a
dedicated SCSS file, it will be easier to maintain them. Some classes
have been added, mainly a class to hide things and a class to replace
the Scriptaculous highlight animation.
Script locale has been replaced by gettext.
@tuleap/dom now supports keys with the full data attribute name (like
`data-project-id`) for getDatasetItemOrThrow(), the goal is to make it
easier to track usage of data attributes without having to do the
conversion to camelCase.

Change-Id: I5fcc6bc78790e7bf9900416b71525438dd98408a