feat: Implements PATCH artidoc/:id/sections
This allows to reorder sections of a document.

To test, have an artidoc with 2 or more sections.
Open the api exlorer and with :id being the document id, execute the
following query (adapt UUIDs to your own):
```
PATCH artidoc/:id/sections
{
  "ids": [
    "01936439-d93c-736f-af80-868d2f184cdf"
  ],
  "direction": "before",
  "compared_to": "01935460-ee9a-7248-886f-740a4ee06f27"
}
```

Note: You can get the UUID of the section to move by clicking on the section
in the TOC, the UUID will be displayed in the URL.

Refresh the artidoc, the section should be reordered.

Part of story #40457: Modify flat document outline

Change-Id: I72d4039ee8963b0d7f20bb28e47a214839ecd3db