feat: Introduce PATCH artidoc/:id/sections
To not clutter the review only the verifications are done in this
contribution. Actual reorder will be done in a dedicated one.

To test, have an artidoc with 2 or more sections.
Open the api exlorer and with :id being the document id, execute the
following query (adapt UUIDs to your own):
```
PATCH artidoc/:id/sections
{
  "ids": [
    "01936439-d93c-736f-af80-868d2f184cdf"
  ],
  "direction": "before",
  "compared_to": "01935460-ee9a-7248-886f-740a4ee06f27"
}
```

Note: You can get the UUID of the section to move by clicking on the section
in the TOC, the UUID will be displayed in the URL.

If everything is ok, then you should get a 404: Not Yet Implemented.

You can test errors:
* `ids` and `compared_to` are the same
* `direction` is unknown
* `ids` contains more than one item
* invalid UUIDS

Implementation note: the handler of the PATCH is not moved yet in the
Adapter/Domain namespace, it will be done later in dedicated contribution. The
route is not fully hexagonal however: 1) most of the business logic of the
move (verifications) is in the Domain, 2) implementing the REST route
right now allows the frontend developpers to use the REST route in
parallel of hexagonalification of the backend.

Part of story #40457: Modify flat document outline

Change-Id: Ic8e87545063aeb1a27921b8a396d39e4a8633db7